Move qt_for_kernel.py from external to lib
After these edits, qt_for_kernel is not really a 3rd party library anymore.
Moreover, I can't even find the original owner of this file -- it seems to be
ipython-specific... so it really ought not be in external.
