+++ IPython/lib/inputhookqt4.py
@@ -21,7 +21,7 @@ import signal
 import threading
 
 from IPython.core.interactiveshell import InteractiveShell
-from IPython.external.qt_for_kernel import QtCore, QtGui
+from IPython.lib.qt_for_kernel import QtCore, QtGui
 from IPython.lib.inputhook import allow_CTRL_C, ignore_CTRL_C, stdin_ready
 
 #-----------------------------------------------------------------------------
