+++ IPython/lib/qt_for_kernel.py
